Clean up JCLI utils and error handling
