+++ src/bin/jcli.rs
@@ -19,8 +19,21 @@ extern crate strfmt;
 
 mod jcli_app;
 
+use std::error::Error;
 use structopt::StructOpt;
 
 fn main() {
-    jcli_app::JCli::from_args().exec()
+    jcli_app::JCli::from_args()
+        .exec()
+        .unwrap_or_else(report_error)
+}
+
+fn report_error(error: Box<Error>) {
+    eprintln!("{}", error);
+    let mut source = error.source();
+    while let Some(sub_error) = source {
+        eprintln!("  |-> {}", sub_error);
+        source = sub_error.source();
+    }
+    std::process::exit(1)
 }
