+++ src/bin/jcli_app/utils/mod.rs
@@ -1,10 +1,5 @@
-#![allow(dead_code)]
-
 mod host_addr;
 pub mod io;
 pub mod key_parser;
-mod segment_parser;
-pub mod serde_with_string;
 
 pub use self::host_addr::HostAddr;
-pub use self::segment_parser::SegmentParser;
